Addition of problem case "sharingcriminalization"
Addition of problem case "I copy, share, download or stream copyright
protected content or would like to and therefore feel criminalized.",
which is only connected to hint 23, and goes in hand with minimal,
additive changes at hints 23 and 80 - just how written down in the pad:
https://piratenpad.de/p/complete-copywrongsEU

This change was done for the english and german version.

Process:
1. Incorporated minimal, additive changes of the hints 23 and 80 in
guides\30c3.csv and guides\30c3_de.csv
2. Incorporated problem case "sharingcriminalization" in de\index.html,
index.html and js\30c3,js .

Order: While doing so, this problem case is placed, whare the problem
case "filesharingconfusion" is, and placed the problem case
"filesharingconfusion" between "cclicenses" and "enforcement".